\name{piecewisesim-package}
\alias{piecewisesim-package}
\alias{piecewisesim}
\docType{package}
\title{
\packageTitle{piecewisesim}
}
\description{
\packageDescription{piecewisesim}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{piecewisesim}
\packageIndices{piecewisesim}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{piecewisesim}

Maintainer: \packageMaintainer{piecewisesim}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
